+++ mysite06/src/main/java/com/poscodx/mysite/service/GuestbookService.java
@@ -17,7 +17,7 @@ public class GuestbookService {
 		return guestbookRepository.findAll();
 	}
 	
-	public Boolean deleteContents(Long no, String password) {
+	public Boolean deleteContents(int no, String password) {
 		return guestbookRepository.deleteByNoAndPassword(no, password);
 	}
